This commit addresses two related issues:
1. In the past, Patch would always read in "old mode" fields in diff
files as being "0". This was due to newlines resetting permissions
read from the diff file. Now old mode data is read from the diff file
properly.

2. If a file being patched had a file permission mode different from
the "old mode" field in a diff, Patch would apply the new mode without
warning about the mis-match. This commit adds a warning. Also, if
--force is not being used, Patch will refuse to change the file permissions
when they don't match the "old mode" field in the diff.